wip: sharing event handlers between sheets
