+++ web/src/components/import/DataSheetView.js
@@ -2,7 +2,6 @@ import React, {useEffect, useState} from 'react';
 import Box from '@mui/material/Box';
 import Button from '@mui/material/Button';
 import Typography from '@mui/material/Typography';
-import {blue, grey, orange, red, yellow} from '@mui/material/colors';
 import CloudDownloadIcon from '@mui/icons-material/CloudDownload';
 import CloudUploadIcon from '@mui/icons-material/CloudUpload';
 import PlaylistAddCheckOutlinedIcon from '@mui/icons-material/PlaylistAddCheckOutlined';
@@ -17,134 +16,29 @@ import LoadingOverlay from '../overlays/LoadingOverlay';
 import AlertDialog from '../ui/AlertDialog';
 import FindReplacePanel from './panel/FindReplacePanel';
 import ValidationPanel from './panel/ValidationPanel';
-import dh from './DataSheetEventHandlers';
+import eh from './DataSheetEventHandlers';
 
-const resetContext = (context) => {
-  context.useOverlay = 'Loading';
-  context.rowData = [];
-  context.rowPos = [];
-  context.highlighted = [];
-  context.putRowIds = [];
-  context.summary = [];
-  context.errorList = {};
-  context.errors = [];
-  context.pendingPasteUndo = [];
-  context.focusedRows = [];
-  context.pasteMode = false;
-};
+// |context| is where all custom properties and helper functions
+// associated with the ag-grid are stored
+// see: https://www.ag-grid.com/javascript-grid/context/
+const context = {
+  useOverlay: 'Loading',
+  rowData: [],
+  highlighted: [],
+  putRowIds: [],
+  undoStack: [],
+  summary: [],
+  errors: [],
+  fullRefresh: false,
 
-const pushUndo = (api, delta) => {
-  const ctx = api.gridOptionsWrapper.gridOptions.context;
-  // |delta| is an array of rowData
-  ctx.undoStack.push(
-    delta.map((d) => {
-      ctx.putRowIds.push(d.id);
-      return {...d};
-    })
-  );
-  return ctx.undoStack.length;
-};
+  // needed by the panels
+  pushUndo: eh.pushUndo,
+  popUndo: eh.popUndo,
 
-const fillRegion = (e, fill) => {
-  const rowData = e.context.rowData;
-  const [cells] = e.api.getCellRanges();
-  const fields = cells.columns.map((col) => col.colId);
-  const delta = [];
-  const startIdx = Math.min(cells.startRow.rowIndex, cells.endRow.rowIndex);
-  const endIdx = Math.max(cells.startRow.rowIndex, cells.endRow.rowIndex);
-  for (let i = startIdx; i < endIdx + 1; i++) {
-    const row = e.api.getDisplayedRowAtIndex(i);
-    const dataIdx = rowData.findIndex((d) => d.id == row.data.id);
-    const data = {...rowData[dataIdx]};
-    delta.push(data);
-    let newData = {};
-    Object.keys(data).forEach(function (key) {
-      newData[key] = data[key];
-    });
-    fields.forEach((key) => {
-      if (key != 'pos') {
-        newData[key] = fill;
-      }
-    });
-    rowData[dataIdx] = newData;
-  }
-  pushUndo(e.api, delta);
-  e.api.setRowData(rowData);
-};
-
-const popUndo = (api) => {
-  const ctx = api.gridOptionsWrapper.gridOptions.context;
-  const deltaSet = ctx.undoStack.pop();
-  let rowData = ctx.rowData;
-  for (const deltaIdx in deltaSet) {
-    const deltaId = deltaSet[deltaIdx].id;
-    ctx.putRowIds.push(deltaId);
-    const rowIdx = rowData.findIndex((d) => d.id === deltaId);
-    if (Object.keys(deltaSet[deltaIdx]).length < 2) {
-      rowData.splice(rowIdx, 1);
-    } else {
-      if (rowIdx < 0) {
-        rowData.push(deltaSet[deltaIdx]);
-      } else {
-        rowData[rowIdx] = deltaSet[deltaIdx];
-      }
-    }
-  }
-  context.fullRefresh = true;
-  api.setRowData(rowData);
-  ctx.rowPos = rowData.map((r) => r.pos).sort((a, b) => a - b);
-  return ctx.undoStack.length;
-};
-
-const chooseCellStyle = (params) => {
-  // Grey-out the first  column containing the row number
-  if (!params.colDef.editable) return {color: grey[800], backgroundColor: grey[50]};
-
-  // Highlight and search results
-  const row = params.context.highlighted[params.rowIndex];
-  if (row && row[params.colDef.field]) return {backgroundColor: yellow[100]};
-
-  // Highlight cell validations
-  const error = params.context.errors.find(
-    (e) => e.rowIds.includes(params.data.id) && (!e.columnNames || e.columnNames.includes(params.colDef.field))
-  );
-
-  switch (error?.levelId) {
-    case 'BLOCKING':
-      return {backgroundColor: red[100]};
-    case 'WARNING':
-      return {backgroundColor: orange[100]};
-    case 'DUPLICATE':
-      if (context.focusedRows?.includes(params.data.id)) {
-        return {backgroundColor: blue[100], fontWeight: 'bold'};
-      } else {
-        return {backgroundColor: blue[100]};
-      }
-    case 'INFO':
-      return {backgroundColor: grey[100]};
-  }
-};
-
-const onSortChanged = (e) => {
-  e.api.refreshCells();
-};
-
-const toolTipValueGetter = (params) => {
-  const error = params.context.errors.find(
-    (e) => e.rowIds.includes(params.data.id) && (!e.columnNames || e.columnNames.includes(params.colDef.field))
-  );
-
-  if (error?.levelId === 'DUPLICATE') {
-    const rowPositions = error.rowIds.map((r) => params.context.rowData.find((d) => d.id === r)?.pos).filter((r) => r);
-    const duplicates = rowPositions.map((r) => params.context.rowPos.indexOf(r) + 1);
-    return duplicates.length > 1 ? 'Rows are duplicated: ' + duplicates.join(', ') : 'Duplicate rows have been removed';
-  }
-
-  return error?.message;
-};
-
-const rowValueGetter = (params) => {
-  return params.context.rowPos ? params.context.rowPos.indexOf(params.data.pos) + 1 : 0;
+  // paste operations must be done row-by-row. build up |pendingPasteUndo|
+  // while in paste mode, then when onPasteEnd is called then call pushUndo
+  pendingPasteUndo: [],
+  pasteMode: false
 };
 
 const defaultSideBar = {
@@ -167,274 +61,9 @@ const defaultSideBar = {
   defaultToolPanel: ''
 };
 
-const generateErrorTree = (rowData, rowPos, errors) => {
-  const tree = {blocking: [], warning: [], info: [], duplicate: []};
-  errors
-    .sort((a, b) => (a.message < b.message ? -1 : a.message > b.message ? 1 : 0))
-    .forEach((e) => {
-      const rows = rowData.filter((r) => e.rowIds.includes(r.id));
-      let summary = [];
-      if (e.columnNames && e.categoryId !== 'SPAN') {
-        const col = e.columnNames[0];
-        summary = rows.reduce((acc, r) => {
-          const rowPosition = rowData.find((d) => d.id === r.id)?.pos;
-          const rowNumber = rowPos.indexOf(rowPosition) + 1;
-          const existingIdx = acc.findIndex((m) => m.columnName === col && m.value === r[col]);
-          if (existingIdx >= 0 && isNaN(parseInt(acc[existingIdx].columnName)))
-            acc[existingIdx] = {
-              columnName: col,
-              value: r[col],
-              rowIds: [...acc[existingIdx].rowIds, r.id],
-              rowNumbers: [...acc[existingIdx].rowNumbers, rowNumber]
-            };
-          else
-            acc.push({columnName: col, value: r[col], rowIds: [r.id], rowNumbers: [rowNumber], isInvertSize: r.isInvertSizing === 'Yes'});
-          return acc;
-        }, []);
-      } else {
-        const rowPositions = e.rowIds.map((r) => rowData.find((d) => d.id === r)?.pos).filter((r) => r);
-        const rowNumbers = rowPositions.map((r) => rowPos.indexOf(r) + 1);
-        summary = [{rowIds: e.rowIds, columnNames: e.columnNames, rowNumbers}];
-      }
-      tree[e.levelId.toLowerCase()].push({key: `err-${e.id}`, message: e.message, count: e.rowIds.length, description: summary});
-    });
-  return tree;
-};
-
-const onPasteStart = (e) => {
-  e.api.gridOptionsWrapper.gridOptions.context.pasteMode = true;
-};
-
-const onCopyRegion = (e) => {
-  e.api.copySelectedRangeToClipboard();
-};
-
-const onClickExcelExport = (api, name, isExtended) => {
-  const columns = [
-    'id',
-    'diver',
-    'buddy',
-    'siteCode',
-    'siteName',
-    'latitude',
-    'longitude',
-    'date',
-    'vis',
-    'direction',
-    'time',
-    'P-Qs',
-    'depth',
-    'method',
-    'block',
-    'code',
-    'species',
-    'commonName',
-    'total',
-    'inverts',
-    ...measurements.map((m) => m.field)
-  ];
-
-  const extendedColumns = [...extendedMeasurements.map((m) => m.field), 'isInvertSizing'];
-  const requiredColumns = isExtended ? [...columns, ...extendedColumns] : columns;
-  const headers = [];
-
-  requiredColumns.forEach((x) => {
-    // Get the row display name from the fields, this is because we turn on skipColumnHeaders so that
-    // we can add empty row, '' is used to force type to string.
-    headers.push({data: {value: '' + api.getColumnDefs().filter((y) => y.field === x)[0].headerName, type: 'String'}});
-  });
-
-  api.exportDataAsExcel({
-    sheetName: 'DATA',
-    author: 'NRMN',
-    columnKeys: requiredColumns,
-    skipColumnHeaders: true,
-    prependContent: [headers, []],
-    fileName: `export_${name}`
-  });
-};
-
-const onCellKeyDown = (e) => {
-  const editingCells = e.api.getEditingCells();
-  if (editingCells.length === 1) {
-    e.api.gridOptionsWrapper.gridOptions.context.navigationKey = e.event.key;
-    if (['ArrowLeft', 'ArrowUp'].includes(e.event.key)) {
-      e.event.preventDefault();
-      e.api.stopEditing();
-      e.api.tabToPreviousCell();
-    }
-    if (['ArrowRight', 'ArrowDown'].includes(e.event.key)) {
-      e.event.preventDefault();
-      e.api.stopEditing();
-      e.api.tabToNextCell();
-    }
-    e.api.gridOptionsWrapper.gridOptions.context.navigationKey = '';
-  }
-};
-
-const onTabToNextCell = (e) => {
-  let context = e.api.gridOptionsWrapper.gridOptions.context;
-  let result;
-
-  if (['ArrowUp', 'ArrowDown'].includes(context.navigationKey) && e.previousCellPosition) {
-    let previousCell = e.previousCellPosition,
-      lastRowIndex = previousCell.rowIndex,
-      nextRowIndex = e.backwards ? lastRowIndex - 1 : lastRowIndex + 1,
-      renderedRowCount = e.api.getModel().getRowCount();
-
-    if (nextRowIndex < 0) nextRowIndex = -1;
-    if (nextRowIndex >= renderedRowCount) nextRowIndex = renderedRowCount - 1;
-
-    result = {
-      rowIndex: nextRowIndex,
-      column: previousCell.column,
-      floating: previousCell.floating
-    };
-  }
-
-  if (['ArrowLeft', 'ArrowRight'].includes(context.navigationKey) && e.nextCellPosition) {
-    result = {
-      rowIndex: e.nextCellPosition.rowIndex,
-      column: e.nextCellPosition.column,
-      floating: e.nextCellPosition.floating
-    };
-  }
-
-  return result;
-};
-
-const getContextMenuItems = (e) => {
-  const [cells] = e.api.getCellRanges();
-  if (!cells) return;
-
-  const colId = cells.startColumn.colId;
-  const row = e.api.getDisplayedRowAtIndex(cells.startRow.rowIndex);
-  const label = row.data[colId];
-
-  let rowData = e.context.rowData;
-  const items = [];
-
-  if (label) {
-    items.push({
-      name: `Fill with '${label}'`,
-      action: () => fillRegion(e, label)
-    });
-  }
-
-  const cloneRow = (clearData) => {
-    const [cells] = e.api.getCellRanges();
-    const row = e.api.getDisplayedRowAtIndex(cells.startRow.rowIndex);
-    const data = rowData.find((d) => d.id == row.data.id);
-    const newId = +new Date().valueOf();
-    const posMap = rowData.map((r) => r.pos).sort((a, b) => a - b);
-    const currentPosIdx = posMap.findIndex((p) => p == data.pos);
-    let newData = {};
-    Object.keys(data).forEach(function (key) {
-      newData[key] = clearData ? '' : data[key];
-    });
-    delete newData.errors;
-    newData.pos = posMap[currentPosIdx + 1] ? posMap[currentPosIdx + 1] - 1 : posMap[currentPosIdx] + 1000;
-    newData.id = newId;
-    pushUndo(e.api, [{id: newId}]);
-    rowData.push(newData);
-    e.api.setRowData(rowData);
-    e.context.rowPos = rowData.map((r) => r.pos).sort((a, b) => a - b);
-    const values = e.api.getRenderedNodes().reduce((acc, field) => acc.concat(field.id.toString()), [newId.toString()]);
-    e.api.setFilterModel({
-      id: {
-        type: 'set',
-        values: values
-      }
-    });
-  };
-
-  const multiRowsSelected = e.api.getSelectedRows().length > 1 || cells.startRow.rowIndex !== cells.endRow.rowIndex;
-  if (!multiRowsSelected) {
-    if (items.length > 0) items.push('separator');
-    items.push({
-      name: 'Delete Row',
-      action: () => deleteRow(e)
-    });
-    items.push({
-      name: 'Clone Row',
-      action: () => cloneRow(false)
-    });
-    items.push({
-      name: 'Insert Row',
-      action: () => cloneRow(true)
-    });
-  } else {
-    if (items.length > 0) items.push('separator');
-    items.push({
-      name: 'Delete Selected Rows',
-      action: () => deleteRow(e)
-    });
-  }
-  return items;
-};
-
-const deleteRow = (e) => {
-  const rowData = e.context.rowData;
-  const [cells] = e.api.getCellRanges();
-  const startIdx = Math.min(cells.startRow.rowIndex, cells.endRow.rowIndex);
-  const endIdx = Math.max(cells.startRow.rowIndex, cells.endRow.rowIndex);
-  const delta = [];
-
-  if (startIdx === endIdx && startIdx === e.node.rowIndex) {
-    e.api.getSelectedRows().forEach(() => {
-      const data = e.node.data;
-      delta.push({...data});
-      rowData.splice(rowData.indexOf(data), 1);
-    });
-  } else if (startIdx === endIdx) {
-    e.api.getSelectedRows().forEach((row) => {
-      const data = rowData.find((d) => d.id === row.id);
-      delta.push({...data});
-      rowData.splice(rowData.indexOf(data), 1);
-    });
-  } else {
-    for (let i = startIdx; i < endIdx + 1; i++) {
-      const row = e.api.getDisplayedRowAtIndex(i);
-      const data = rowData.find((d) => d.id === row.data.id);
-      delta.push({...data});
-      rowData.splice(rowData.indexOf(data), 1);
-    }
-  }
-  pushUndo(e.api, delta);
-  e.api.setRowData(rowData);
-  e.context.rowPos = rowData.map((r) => r.pos).sort((a, b) => a - b);
-  e.api.refreshCells();
-};
-
-const overrideKeyboardEvents = (e) => {
-  if (e.event.key === 'Delete') {
-    if (e.event.type === 'keydown') onClearRegion(e);
-    return true;
-  }
-
-  if (e.event.ctrlKey || e.event.metaKey) {
-    if (e.event.key === 'c' && e.event.type === 'keydown') {
-      onCopyRegion(e);
-      return true;
-    }
-    if (e.event.key === 'x' && e.event.type === 'keydown') {
-      onCutRegion(e);
-      return true;
-    }
-    return false;
-  }
-};
-
-const onClearRegion = (e) => fillRegion(e, '');
-
-const onCutRegion = (e) => {
-  onCopyRegion(e);
-  onClearRegion(e);
-};
-
 const reload = (api, id, completion, isAdmin) => {
   const context = api.gridOptionsWrapper.gridOptions.context;
-  resetContext(context);
+  eh.resetContext();
   context.isAdmin = isAdmin;
   getDataJob(id).then((res) => {
     const job = {
@@ -461,69 +90,6 @@ const reload = (api, id, completion, isAdmin) => {
   });
 };
 
-const handlePasteEnd = (e) => {
-  const ctx = e.api.gridOptionsWrapper.gridOptions.context;
-  ctx.pasteMode = false;
-  let oldRows = [];
-  Array.from(new Set(ctx.pendingPasteUndo.map((u) => u.id))).forEach((id) => {
-    let oldRow = {};
-    let rowData = ctx.rowData;
-    const newRow = rowData.find((r) => r.id === id);
-    Object.keys(newRow).forEach(function (key) {
-      oldRow[key] = newRow[key];
-    });
-    ctx.pendingPasteUndo
-      .filter((u) => u.id === id)
-      .forEach((p) => {
-        const field = p.field;
-        oldRow[field] = p.value;
-      });
-    oldRows.push(oldRow);
-  });
-  ctx.pendingPasteUndo = [];
-  return ctx.pushUndo(e.api, [...oldRows]);
-};
-
-const handleCellValueChanged = (e) => {
-  if (e.context.pasteMode)
-    e.context.pendingPasteUndo.push({id: e.data.id, field: e.colDef.field, value: e.oldValue});
-
-  return e.context.undoStack.length;
-};
-
-const handleUndo = (e) => {
-  popUndo(e.api);
-  e.api.refreshCells();
-};
-
-const handleCellEditingStopped = (e) => {
-  if (e.oldValue === e.newValue) return;
-  const row = {...e.data};
-  row[e.column.colId] = e.oldValue;
-  return pushUndo(e.api, [row]);
-};
-
-// |context| is where all custom properties and helper functions
-// associated with the ag-grid are stored
-// see: https://www.ag-grid.com/javascript-grid/context/
-const context = {
-  useOverlay: 'Loading',
-  rowData: [],
-  highlighted: [],
-  putRowIds: [],
-  undoStack: [],
-  summary: [],
-  errors: [],
-  fullRefresh: false,
-  pushUndo: pushUndo,
-  popUndo: popUndo,
-
-  // paste operations must be done row-by-row. build up |pendingPasteUndo|
-  // while in paste mode, then when onPasteEnd is called then call pushUndo
-  pendingPasteUndo: [],
-  pasteMode: false
-};
-
 const IngestState = Object.freeze({Loading: 0, Edited: 1, Valid: 2, ConfirmSubmit: 3});
 
 const DataSheetView = ({onIngest, isAdmin}) => {
@@ -540,14 +106,14 @@ const DataSheetView = ({onIngest, isAdmin}) => {
       if (event.ctrlKey && event.key === 'z') {
         event.preventDefault();
         event.stopPropagation();
-        onUndo({api: gridApi});
+        eh.handleUndo({api: gridApi});
       }
     };
     document.body.addEventListener('keydown', undoKeyboardHandler);
     return () => {
       document.body.removeEventListener('keydown', undoKeyboardHandler);
     };
-  });
+  },[gridApi]);
 
   useEffect(() => {
     if (gridApi && state === IngestState.Loading) gridApi.showLoadingOverlay();
@@ -574,7 +140,7 @@ const DataSheetView = ({onIngest, isAdmin}) => {
       delete result.data.errors;
       delete result.data.job;
       context.summary = result.data;
-      context.errorList = generateErrorTree(context.rowData, context.rowPos, context.errors);
+      context.errorList = eh.generateErrorTree(context.rowData, context.rowPos, context.errors);
 
       setState(context.errors.some((e) => e.levelId === 'BLOCKING') ? IngestState.Edited : IngestState.Valid);
 
@@ -644,20 +210,16 @@ const DataSheetView = ({onIngest, isAdmin}) => {
   };
 
   const onCellValueChanged = (e) => {
-    setUndoSize(handleCellValueChanged(e));
+    setUndoSize(eh.handleCellValueChanged(e));
   };
 
   const onPasteEnd = (e) => {
-    setUndoSize(handlePasteEnd(e));
-  };
-
-  const onUndo = (e) => {
-    if (undoSize < 1) return;
-    handleUndo(e);
+    setUndoSize(eh.handlePasteEnd(e));
   };
 
   const onCellEditingStopped = (e) => {
-    setUndoSize(handleCellEditingStopped(e));
+    const undos = eh.handleCellEditingStopped(e);
+    setUndoSize(undos);
     setState(IngestState.Edited);
   };
 
@@ -712,7 +274,7 @@ const DataSheetView = ({onIngest, isAdmin}) => {
               } `}</Typography>
             </Box>
             <Box m={1} ml={0}>
-              <Button variant="outlined" disabled={undoSize < 1} startIcon={<UndoIcon />} onClick={() => onUndo({api: gridApi})}>
+              <Button variant="outlined" disabled={undoSize < 1} startIcon={<UndoIcon />} onClick={() => eh.handleUndo({api: gridApi})}>
                 Undo
               </Button>
             </Box>
@@ -724,7 +286,7 @@ const DataSheetView = ({onIngest, isAdmin}) => {
             <Box m={1} ml={0}>
               <Button
                 variant="outlined"
-                onClick={() => onClickExcelExport(gridApi, job.reference, job.isExtendedSize)}
+                onClick={() => eh.onClickExcelExport(gridApi, job.reference, job.isExtendedSize)}
                 startIcon={<CloudDownloadIcon />}
               >
                 Export
@@ -772,10 +334,10 @@ const DataSheetView = ({onIngest, isAdmin}) => {
               filter: true,
               floatingFilter: true,
               suppressMenu: true,
-              suppressKeyboardEvent: overrideKeyboardEvents,
-              cellStyle: chooseCellStyle,
+              suppressKeyboardEvent: eh.overrideKeyboardEvents,
+              cellStyle: eh.chooseCellStyle,
               enableCellChangeFlash: true,
-              tooltipValueGetter: toolTipValueGetter,
+              tooltipValueGetter: eh.toolTipValueGetter,
               valueParser: ({newValue}) => (newValue ? newValue.trim() : '')
             }}
             rowHeight={20}
@@ -784,16 +346,16 @@ const DataSheetView = ({onIngest, isAdmin}) => {
             enableRangeSelection={true}
             animateRows={true}
             enableRangeHandle={true}
-            onCellKeyDown={onCellKeyDown}
-            onPasteStart={onPasteStart}
+            onCellKeyDown={eh.onCellKeyDown}
+            onPasteStart={eh.onPasteStart}
             onPasteEnd={onPasteEnd}
-            tabToNextCell={onTabToNextCell}
+            tabToNextCell={eh.onTabToNextCell}
             onCellValueChanged={onCellValueChanged}
-            onSortChanged={onSortChanged}
+            onSortChanged={eh.onSortChanged}
             onFilterChanged={onFilterChanged}
             onRowDataUpdated={onRowDataUpdated}
             fillHandleDirection="y"
-            getContextMenuItems={getContextMenuItems}
+            getContextMenuItems={(e) => eh.getContextMenuItems(e,eh)}
             undoRedoCellEditing={false}
             onCellEditingStopped={onCellEditingStopped}
             components={{
@@ -814,7 +376,7 @@ const DataSheetView = ({onIngest, isAdmin}) => {
               headerName=""
               suppressMovable={true}
               editable={false}
-              valueGetter={rowValueGetter}
+              valueGetter={eh.rowValueGetter}
               minWidth={40}
               enableCellChangeFlash={false}
               filter={false}
@@ -826,7 +388,7 @@ const DataSheetView = ({onIngest, isAdmin}) => {
             <AgGridColumn field="siteName" headerName="Site Name" minWidth={160} />
             <AgGridColumn field="latitude" headerName="Latitude" />
             <AgGridColumn field="longitude" headerName="Longitude" />
-            <AgGridColumn field="date" headerName="Date" rowGroup={false} enableRowGroup={true} comparator={dh.dateComparator} />
+            <AgGridColumn field="date" headerName="Date" rowGroup={false} enableRowGroup={true} comparator={eh.dateComparator} />
             <AgGridColumn field="vis" headerName="Vis" />
             <AgGridColumn field="direction" headerName="Direction" />
             <AgGridColumn field="time" headerName="Time" />
